+++ packages/core/form/examples/00-signup-form.js
@@ -1,0 +1,97 @@
+// @flow
+import React from 'react';
+import TextField from '@atlaskit/textfield';
+import Button, { ButtonGroup } from '@atlaskit/button';
+import { Checkbox } from '@atlaskit/checkbox';
+import Form, {
+  CheckboxField,
+  Field,
+  FormFooter,
+  HelperMessage,
+  ErrorMessage,
+  ValidMessage,
+} from '../src';
+
+export default () => (
+  <div
+    style={{
+      display: 'flex',
+      width: '400px',
+      margin: '0 auto',
+      flexDirection: 'column',
+    }}
+  >
+    <Form
+      onSubmit={data => {
+        console.log('form data', data);
+        return new Promise(resolve => setTimeout(resolve, 2000)).then(() =>
+          data.username === 'error' ? { username: 'IN_USE' } : undefined,
+        );
+      }}
+    >
+      {({ formProps, submitting }) => (
+        <form {...formProps}>
+          <Field name="username" label="User name" isRequired defaultValue="">
+            {({ fieldProps, error }) => (
+              <>
+                <TextField autoComplete="off" {...fieldProps} />
+                {!error && (
+                  <HelperMessage fieldId={fieldProps.id}>
+                    You can use letters, numbers & periods.
+                  </HelperMessage>
+                )}
+                {error && (
+                  <ErrorMessage fieldId={fieldProps.id}>
+                    This user name is already in use, try another one.
+                  </ErrorMessage>
+                )}
+              </>
+            )}
+          </Field>
+          <Field
+            name="password"
+            label="Password"
+            defaultValue=""
+            isRequired
+            validate={value => (value.length < 8 ? 'TOO_SHORT' : undefined)}
+          >
+            {({ fieldProps, error, meta }) => (
+              <>
+                <TextField type="password" {...fieldProps} />
+                {!error && !meta.valid && (
+                  <HelperMessage fieldId={fieldProps.id}>
+                    Use 8 or more characters with a mix of letters, numbers &
+                    symbols.
+                  </HelperMessage>
+                )}
+                {error && (
+                  <ErrorMessage fieldId={fieldProps.id}>
+                    Password needs to be more than 8 characters.
+                  </ErrorMessage>
+                )}
+                {meta.valid && (
+                  <ValidMessage fieldId={fieldProps.id}>
+                    Awesome password!
+                  </ValidMessage>
+                )}
+              </>
+            )}
+          </Field>
+          <CheckboxField name="remember" label="Remember me" defaultIsChecked>
+            {({ fieldProps }) => (
+              <Checkbox {...fieldProps} label="Always sign in on this device" />
+            )}
+          </CheckboxField>
+          <FormFooter>
+            <ButtonGroup>
+              <Button appearance="subtle">Cancel</Button>
+              <Button type="submit" appearance="primary" isLoading={submitting}>
+                Sign up
+              </Button>
+            </ButtonGroup>
+          </FormFooter>
+        </form>
+      )}
+    </Form>
+  </div>
+);