+++ packages/editor/editor-core/src/__tests__/unit/plugins/media/ui/MediaSingleEdit.tsx
@@ -105,11 +105,7 @@ describe('@atlaskit/editor-core/ui/MediaSingleEdit', () => {
       'mediaSingle',
     );
 
-    expect(
-      mediaSingleEdit.find(ToolbarButton).findWhere(child => {
-        return child.type() === ToolbarButton && child.prop('disabled');
-      }).length,
-    ).toEqual(5);
+    expect(mediaSingleEdit.find(ToolbarButton).length).toEqual(1);
     mediaSingleEdit.unmount();
   });
 
@@ -148,11 +144,47 @@ describe('@atlaskit/editor-core/ui/MediaSingleEdit', () => {
     expect(editorView.state.selection.$from.node().type.name).toEqual(
       'mediaSingle',
     );
-    expect(
-      mediaSingleEdit.find(ToolbarButton).findWhere(child => {
-        return child.type() === ToolbarButton && child.prop('disabled');
-      }).length,
-    ).toEqual(5);
+    expect(mediaSingleEdit.find(ToolbarButton).length).toEqual(1);
+    mediaSingleEdit.unmount();
+  });
+
+  it('should not have breakout options if resizing is allowed', () => {
+    const { editorView } = editor(
+      doc(
+        layoutSection(
+          layoutColumn({ width: 50 })(p('')),
+          layoutColumn({ width: 50 })(
+            mediaSingle({ layout: 'center' })(
+              media({
+                id: temporaryFileId,
+                __key: temporaryFileId,
+                type: 'file',
+                collection: testCollectionName,
+                __fileMimeType: 'image/png',
+                width: 100,
+                height: 100,
+              })(),
+            ),
+          ),
+        ),
+      ),
+    );
+    editorView.focus();
+    const pluginState = pluginKey.getState(editorView.state);
+
+    const mediaSingleEdit = mountWithIntl(
+      <MediaSingleEdit
+        pluginState={pluginState}
+        target={pluginState.element}
+        allowBreakout={true}
+        allowLayout={false}
+        allowResizing={true}
+      />,
+    );
+    expect(editorView.state.selection.$from.node().type.name).toEqual(
+      'mediaSingle',
+    );
+    expect(mediaSingleEdit.find(ToolbarButton).length).toEqual(1);
     mediaSingleEdit.unmount();
   });
 });