+++ packages/editor/editor-mobile-bridge/src/__tests__/integration/_utils.ts
@@ -6,6 +6,19 @@ declare global {
   }
 }
 
+// FIXME Ideally these would be mobile browsers
+// Safari & Chrome should suffice for now.
+export const skipBrowsers: any = ['ie', 'firefox', 'edge'];
+
+export const navigateOrClear = async (browser, path) => {
+  const currentUrl = browser.url();
+  if (currentUrl === path) {
+    await clearEditor(browser);
+  } else {
+    await browser.goto(path);
+  }
+};
+
 export const getDocFromElement = el => el.pmViewDesc.node.toJSON();
 export const editable = '.ProseMirror';
 
@@ -21,10 +34,9 @@ export const renderer = {
   placeholder: '#examples', // FIXME lets add something better to renderer
 };
 
-export const callNativeBridgeFn = async (browser, bridgeFn, ...args) => {
+export const callNativeBridge = async (browser, bridgeFn, ...args) => {
   return await browser.browser.execute(
-    (bridgeFn, args, done) => {
-      // browser context - you may not access client or console
+    (bridgeFn, args) => {
       if (window.bridge && window.bridge[bridgeFn]) {
         window.bridge[bridgeFn].apply(window.bridge, args);
       }
@@ -32,6 +44,25 @@ export const callNativeBridgeFn = async (browser, bridgeFn, ...args) => {
     bridgeFn,
     args || [],
   );
+};
+
+export const getBridgeOutput = async (browser, bridge, bridgeFn) => {
+  const logs = await browser.browser.execute(
+    (bridge, bridgeFn) => {
+      // @ts-ignore
+      let logs = window.logBridge;
+
+      if (logs[`${bridge}:${bridgeFn}`]) {
+        return logs[`${bridge}:${bridgeFn}`];
+      }
+
+      return logs;
+    },
+    bridge,
+    bridgeFn,
+  );
+
+  return logs.value;
 };
 
 export const clearEditor = async browser => {