+++ packages/editor/editor-mobile-bridge/src/editor/web-to-native/ios-impl.ts
@@ -111,4 +111,6 @@ export default class IosBridge implements NativeBridge {
   ) {
     sendToBridge(bridge, event, ...args);
   }
+
+  updateTextColor() {}
 }