+++ packages/editor/renderer/src/ui/Renderer/style.tsx
@@ -78,7 +78,7 @@ const fullPageStyles = ({
 };
 
 // prettier-ignore
-export const Wrapper = styled.div<Props & HTMLAttributes<{}>>`
+export const Wrapper = styled.div < Props & HTMLAttributes < {} >> `
   ${fullPageStyles}
 
   font-size: ${editorFontSize}px;
@@ -203,18 +203,16 @@ export const Wrapper = styled.div<Props & HTMLAttributes<{}>>`
     margin: ${gridSize() * 3}px 0;
   }
 
-  & div > .media-wrapped + .media-wrapped + *:not(.media-wrapped) {
-    clear: both;
-  }
-
-  & .media-wrapped + div:not(.media-wrapped) {
+  .media-single.media-wrapped + .media-single:not(.media-wrapped) {
     clear: both;
   }
 
   & .CodeBlock,
   & blockquote,
   & hr,
-  & > div > div:not(.media-wrapped) {
+  & > div > div:not(.media-wrapped),
+  .media-single.media-wrapped + .media-wrapped + *:not(.media-wrapped),
+  .media-single.media-wrapped + div:not(.media-wrapped) {
     clear: both;
   }
 
@@ -229,8 +227,8 @@ export const Wrapper = styled.div<Props & HTMLAttributes<{}>>`
     }
   }
 
-  ${mediaSingleSharedStyle} & .wrap-left + .wrap-right,
-  & .wrap-right + .wrap-left {
+  ${mediaSingleSharedStyle} & 
+  div[class^='image-wrap-'] + div[class^='image-wrap-'] {
     margin-left: 0;
     margin-right: 0;
   }