+++ packages/media/media-picker/src/service/__tests__/unit/newUploadService.spec.ts
@@ -14,10 +14,12 @@ import {
   FileState,
   MediaItem,
 } from '@atlaskit/media-core';
-import { fakeContext, nextTick } from '@atlaskit/media-test-helpers';
+import { TouchedFiles } from '@atlaskit/media-store';
+import * as uuid from 'uuid';
+import { asMock, fakeContext } from '@atlaskit/media-test-helpers';
 import { Observable } from 'rxjs/Observable';
 import { Subscriber } from 'rxjs';
-import { NewUploadServiceImpl, UpfrontIds } from '../../newUploadServiceImpl';
+import { NewUploadServiceImpl } from '../../newUploadServiceImpl';
 import { MediaFile, UploadParams } from '../../..';
 import * as getPreviewModule from '../../../util/getPreviewFromBlob';
 import * as getPreviewFromImage from '../../../util/getPreviewFromImage';
@@ -34,11 +36,10 @@ describe('UploadService', () => {
   const clientId = 'some-client-id';
   const token = 'some-token';
   const upfrontId = Promise.resolve('1');
-  const authProvider = jest.fn(() =>
-    Promise.resolve<Auth>({ clientId, token, baseUrl }),
-  );
+  let authProvider: AuthProvider;
   const usersClientId = 'some-users-collection-client-id';
   const usersToken = 'some-users-collection-client-id';
+  const previewObject: Preview = { someImagePreview: true } as any;
   const userAuthProvider: AuthProvider = () =>
     Promise.resolve({
       clientId: usersClientId,
@@ -56,9 +57,34 @@ describe('UploadService', () => {
     tenantUploadParams: UploadParams = { collection: '' },
     shouldCopyFileToRecents: boolean = true,
   ) => {
+    const touchedFiles: TouchedFiles = {
+      created: [
+        {
+          fileId: 'uuid1',
+          uploadId: 'some-upload-id-uuid1',
+        },
+        {
+          fileId: 'uuid2',
+          uploadId: 'some-upload-id-uuid2',
+        },
+        {
+          fileId: 'uuid3',
+          uploadId: 'some-upload-id-uuid3',
+        },
+        {
+          fileId: 'uuid4',
+          uploadId: 'some-upload-id-uuid4',
+        },
+      ],
+    };
+    jest.spyOn(context.file, 'touchFiles').mockResolvedValue(touchedFiles);
     jest.spyOn(context.file, 'upload').mockReturnValue({
       subscribe() {},
     });
+
+    (getPreviewFromImage.getPreviewFromImage as any).mockReturnValue(
+      Promise.resolve(previewObject),
+    );
 
     const uploadService = new NewUploadServiceImpl(
       context,
@@ -79,6 +105,9 @@ describe('UploadService', () => {
         .spyOn((uploadService as any).userMediaStore, 'createFile')
         .mockResolvedValue({ data: { id: 'some-new-user-file-id' } });
       const userContext: Context = (uploadService as any)['userContext'];
+      jest
+        .spyOn(userContext.file, 'touchFiles')
+        .mockResolvedValue(touchedFiles);
       const userContextUpload = jest.spyOn(userContext.file, 'upload');
       userContextUpload.mockReturnValue({
         subscribe() {},
@@ -91,11 +120,24 @@ describe('UploadService', () => {
   };
 
   beforeEach(() => {
+    authProvider = jest.fn(() =>
+      Promise.resolve<Auth>({ clientId, token, baseUrl }),
+    );
     fileStreamCacheSpy.mockReset();
     (getPreviewModule.getPreviewFromBlob as any).mockReset();
     (getPreviewModule.getPreviewFromBlob as any).mockReturnValue(
       Promise.resolve(),
     );
+    jest
+      .spyOn(uuid, 'v4')
+      .mockReturnValueOnce('uuid1')
+      .mockReturnValueOnce('uuid2')
+      .mockReturnValueOnce('uuid3')
+      .mockReturnValueOnce('uuid4');
+  });
+
+  afterEach(() => {
+    jest.resetAllMocks();
   });
 
   describe('setUploadParams', () => {
@@ -153,10 +195,6 @@ describe('UploadService', () => {
 
       const callback = jest.fn();
       uploadService.on('file-preview-update', callback);
-      const previewObject: Preview = { someImagePreview: true } as any;
-      (getPreviewFromImage.getPreviewFromImage as any).mockReturnValue(
-        Promise.resolve(previewObject),
-      );
 
       uploadService.addFiles([file as File]);
       await filesAddedPromise;
@@ -262,16 +300,11 @@ describe('UploadService', () => {
         name: 'some-other-filename',
         type: 'image/png',
       } as any;
-      const context = getContext();
-      const uploadFile = jest.fn().mockReturnValue({
-        subscribe() {},
-      });
-      (context as any).file = { upload: uploadFile };
-      const { uploadService } = setup(context, {
+      const { context, uploadService } = setup(undefined, {
         collection: 'some-collection',
       });
       uploadService.addFiles([file, file2]);
-      expect(uploadFile).toHaveBeenCalledTimes(2);
+      expect(context.file.upload).toHaveBeenCalledTimes(2);
       const expectedUploadableFile2: UploadableFile = {
         collection: 'some-collection',
         content: file2,
@@ -284,8 +317,12 @@ describe('UploadService', () => {
         name: 'some-filename',
         mimeType: 'video/mp4',
       };
-      expect(uploadFile.mock.calls[0][0]).toEqual(expectedUploadableFile1);
-      expect(uploadFile.mock.calls[1][0]).toEqual(expectedUploadableFile2);
+      expect(asMock(context.file.upload).mock.calls[0][0]).toEqual(
+        expectedUploadableFile1,
+      );
+      expect(asMock(context.file.upload).mock.calls[1][0]).toEqual(
+        expectedUploadableFile2,
+      );
     });
 
     it.skip('should emit file-converting when uploadFile resolves', async () => {
@@ -854,82 +891,11 @@ describe('UploadService', () => {
       const { uploadService } = setup(undefined, {
         collection: 'some-collection',
       });
-      const upfrontIds: UpfrontIds = {
-        upfrontId: Promise.resolve('1234'),
-      };
-
-      (uploadService as any).getUpfrontIds = jest
-        .fn()
-        .mockReturnValue(upfrontIds);
       uploadService.addFiles([file]);
 
-      await nextTick();
-
       expect(fileStreamCacheSpy).toHaveBeenCalledTimes(2);
-      expect(fileStreamCacheSpy.mock.calls[0][0]).toBe('1234');
-      expect(fileStreamCacheSpy.mock.calls[1][0]).toBe('1234-some-collection');
-    });
-  });
-
-  describe('getUpfrontIds()', () => {
-    let upfrontIds: UpfrontIds;
-    let createFileSpy: jest.SpyInstance<any>;
-
-    describe('when should copy file to recents', () => {
-      beforeEach(async () => {
-        const context = getContext({ userAuthProvider });
-        const { uploadService } = setup(
-          context,
-          { collection: 'tenant-collection-name' },
-          false,
-        );
-        const fileState: FileState = {
-          status: 'processing',
-          id: 'some-user-id',
-          name: 'some-name',
-          occurrenceKey: 'some-user-occurrence-key',
-          size: 42,
-          mediaType: 'image',
-          mimeType: 'some-mime-type',
-        };
-        const observable = new Observable<FileState>(observer => {
-          setImmediate(() => {
-            observer.next(fileState);
-          });
-        });
-
-        createFileSpy = jest
-          .spyOn((uploadService as any).tenantMediaStore, 'createFile')
-          .mockResolvedValue({
-            data: {
-              id: '12',
-            },
-          });
-
-        upfrontIds = await (uploadService as any).getUpfrontIds(
-          observable,
-          'tenant-occurrence-key',
-        );
-      });
-
-      it('should return upfrontId created with createFile API', async () => {
-        expect(createFileSpy).toHaveBeenCalledTimes(1);
-        expect(createFileSpy).lastCalledWith({
-          collection: 'tenant-collection-name',
-          occurrenceKey: 'tenant-occurrence-key',
-        });
-        expect(await upfrontIds.upfrontId).toEqual('12');
-      });
-
-      it('should return userUpfrontId taken from observable', async () => {
-        expect(await upfrontIds.userUpfrontId).toEqual('some-user-id');
-      });
-
-      it('should return userOccurrenceKey taken from observable', async () => {
-        expect(await upfrontIds.userOccurrenceKey).toEqual(
-          'some-user-occurrence-key',
-        );
-      });
+      expect(fileStreamCacheSpy.mock.calls[0][0]).toBe('uuid1');
+      expect(fileStreamCacheSpy.mock.calls[1][0]).toBe('uuid1-some-collection');
     });
   });
 });