+++ packages/people-and-teams/focused-task-close-account/src/components/DeleteUserOverviewScreen/styled.ts
@@ -37,6 +37,9 @@ export const MainInformationList = styled.ul`
       font-weight: 600;
     }
   }
+  p + ul {
+    margin-top: ${gridSizeTimes(1.5)}px;
+  }
 `;
 
 export const IconHoverWrapper = styled.span`