+++ packages/people-and-teams/focused-task-close-account/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/focused-task-close-account
 
+## 0.4.0
+- [minor] [81299f6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/81299f6):
+
+  - Copy changes along with addition on new prop, isUserDeactivated
+
 ## 0.3.4
 - Updated dependencies [58b84fa](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/58b84fa):
   - @atlaskit/avatar@14.1.7